+++ axelor-production/src/main/java/com/axelor/apps/production/module/ProductionModule.java
@@ -55,6 +55,8 @@ import com.axelor.apps.production.service.ProductionOrderService;
 import com.axelor.apps.production.service.ProductionOrderServiceImpl;
 import com.axelor.apps.production.service.ProductionOrderWizardService;
 import com.axelor.apps.production.service.ProductionOrderWizardServiceImpl;
+import com.axelor.apps.production.service.RawMaterialRequirementService;
+import com.axelor.apps.production.service.RawMaterialRequirementServiceImpl;
 import com.axelor.apps.production.service.SaleOrderWorkflowServiceProductionImpl;
 import com.axelor.apps.production.service.StockRulesServiceProductionImpl;
 import com.axelor.apps.production.service.app.AppProductionService;
@@ -100,5 +102,6 @@ public class ProductionModule extends AxelorModule {
     bind(StockMoveLineStockRepository.class).to(StockMoveLineProductionRepository.class);
     bind(ProdProcessLineService.class).to(ProdProcessLineServiceImpl.class);
     bind(ProdProductRepository.class).to(ProdProductProductionRepository.class);
+    bind(RawMaterialRequirementService.class).to(RawMaterialRequirementServiceImpl.class);
   }
 }