+++ axelor-production/src/main/java/com/axelor/apps/production/service/RawMaterialRequirementService.java
@@ -15,13 +15,18 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.axelor.apps.production.report;
+package com.axelor.apps.production.service;
 
-public interface IReport {
+import com.axelor.apps.production.db.RawMaterialRequirement;
+import com.axelor.exception.AxelorException;
 
-  public static final String MANUF_ORDER = "ManufOrder.rptdesign";
-  public static final String OPERATION_ORDER = "OperationOrder.rptdesign";
-  public static final String PROD_PROCESS = "ProdProcess.rptdesign";
-  public static final String BILL_OF_MATERIAL = "BillOfMaterial.rptdesign";
-  public static final String RAW_MATERIAL_REQUIREMENT = "RawMaterialRequirement.rptdesign";
+public interface RawMaterialRequirementService {
+
+  /**
+   * Print the raw material requirement report.
+   *
+   * @param rawMaterialRequirement the user defined parameter of the report.
+   * @return URL to the printed report.
+   */
+  String print(RawMaterialRequirement rawMaterialRequirement) throws AxelorException;
 }