+++ CHANGELOG.md
@@ -50,6 +50,7 @@
 - Business Project : Enhancement of task and project view.
 - Project : Change invoicing project view.
 - Stock Move : Add Wizard for tracking number
+- Production: new report showing raw material requirements.
 
 ## Improvements
 - Move project resource menu entry to project configuration menu.